refactor: split flat storage into multiple files and rename some structs (#8664)
Part of #8577

This includes 2 parts:
* split single-file `near_store::flat_state` module into multiple files under `flat` directory
* rename some structs to better reflect the responsibilities

Please note that this does not introduce any behaviour or API changes, it simply moves the code around.
