+++ core/store/src/trie/trie_tests.rs
@@ -77,8 +77,11 @@ where
     print!("Test touches {} nodes, expected result {:?}...", size, expected);
     for i in 0..(size + 1) {
         let storage = IncompletePartialStorage::new(storage.clone(), i);
-        let new_trie =
-            Trie { storage: Box::new(storage), root: *trie.get_root(), flat_state: None };
+        let new_trie = Trie {
+            storage: Box::new(storage),
+            root: *trie.get_root(),
+            flat_storage_chunk_view: None,
+        };
         let expected_result =
             if i < size { Err(&StorageError::TrieNodeMissing) } else { Ok(&expected) };
         assert_eq!(test(Rc::new(new_trie)).as_ref(), expected_result);
