+++ genesis-tools/genesis-populate/src/lib.rs
@@ -182,7 +182,7 @@ impl GenesisBuilder {
         let mut store_update = tries.store_update();
         let root = tries.apply_all(&trie_changes, shard_uid, &mut store_update);
         #[cfg(feature = "protocol_feature_flat_state")]
-        near_store::FlatStateDelta::from_state_changes(&state_changes)
+        near_store::flat::FlatStateDelta::from_state_changes(&state_changes)
             .apply_to_flat_state(&mut store_update);
         // silence unused variable warning when protocol_feature_flat_state feature is disabled
         drop(state_changes);
