+++ runtime/runtime/src/genesis.rs
@@ -15,7 +15,7 @@ use near_primitives::{
     types::{AccountId, Balance, MerkleHash, ShardId, StateChangeCause, StateRoot},
 };
 #[cfg(feature = "protocol_feature_flat_state")]
-use near_store::flat_state::FlatStateDelta;
+use near_store::flat::FlatStateDelta;
 use near_store::{
     get_account, get_received_data, set, set_access_key, set_account, set_code,
     set_postponed_receipt, set_received_data, ShardTries, TrieUpdate,
