+++ runtime/runtime/src/prefetch.rs
@@ -310,8 +310,8 @@ mod tests {
         let shard_uids = vec![ShardUId::single_shard()];
         let trie_config = TrieConfig { enable_receipt_prefetching: true, ..TrieConfig::default() };
         let store = create_test_store();
-        let flat_storage_factory = near_store::flat_state::FlatStateFactory::new(store.clone());
-        let tries = ShardTries::new(store, trie_config, &shard_uids, flat_storage_factory);
+        let flat_storage_manager = near_store::flat::FlatStorageManager::new(store.clone());
+        let tries = ShardTries::new(store, trie_config, &shard_uids, flat_storage_manager);
 
         let mut kvs = vec![];
