+++ tools/flat-storage/src/commands.rs
@@ -6,10 +6,8 @@ use near_chain::{
 };
 use near_epoch_manager::EpochManagerAdapter;
 use near_primitives::{state::ValueRef, trie_key::trie_key_parsers::parse_account_id_from_raw_key};
-use near_store::{
-    flat_state::store_helper::{self, get_flat_head, get_flat_storage_creation_status},
-    Mode, NodeStorage, ShardUId, Store, StoreOpener,
-};
+use near_store::flat::store_helper::{self, get_flat_head, get_flat_storage_creation_status};
+use near_store::{Mode, NodeStorage, ShardUId, Store, StoreOpener};
 use nearcore::{load_config, NearConfig, NightshadeRuntime};
 use std::{path::PathBuf, sync::Arc, time::Duration};
 use tqdm::tqdm;
@@ -128,14 +126,13 @@ impl FlatStorageCommand {
                 let tip = rw_chain_store.final_head().unwrap();
 
                 // TODO: there should be a method that 'loads' the current flat storage state based on Storage.
-                rw_hot_runtime.create_flat_storage_state_for_shard(
+                rw_hot_runtime.create_flat_storage_for_shard(
                     reset_cmd.shard_id,
                     tip.height,
                     &rw_chain_store,
                 );
 
-                rw_hot_runtime
-                    .remove_flat_storage_state_for_shard(reset_cmd.shard_id, &tip.epoch_id)?;
+                rw_hot_runtime.remove_flat_storage_for_shard(reset_cmd.shard_id, &tip.epoch_id)?;
             }
             SubCommand::Init(init_cmd) => {
                 let (_, rw_hot_runtime, rw_chain_store, rw_hot_store) = Self::get_db(
@@ -198,7 +195,7 @@ impl FlatStorageCommand {
                 println!("Verifying using the {:?} as state_root", state_root);
                 let tip = chain_store.final_head().unwrap();
 
-                hot_runtime.create_flat_storage_state_for_shard(
+                hot_runtime.create_flat_storage_for_shard(
                     verify_cmd.shard_id,
                     tip.height,
                     &chain_store,
